Add a runner class
The current cozy-desktop has a bug when stopping and restarting the
synchronization. https://github.com/cozy-labs/cozy-desktop/pull/1795
aims to fix this issue, but using Promises look a bit messy and quite
hard to get it right to my eyes. I've used this as an exercise for
Crystal. I have no tests, so there are maybe some bugs, but I'm pleased
by how it reads.