Use our own stylesheet
Using cozy-ui (or more generally a CSS framework) can be a good idea,
but it makes modifying the HTML more often (and the CSS less often).
As the views are compiled, and restarting the web server takes something
like 5 seconds on my computer, it makes it painful to change styles via
modifying the HTML. On the other ground, when a public file (CSS or JS)
is modified, we can see the change but refreshing the browser without
having to restart the web server first. So, I prefer to write CSS in
this context.